Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pandas==2.2.3
sqlalchemy==2.0.36
openpyxl==3.1.5
openai==1.54.3
requests==2.32.3
pydantic==2.9.2